size_t

+++ webrtc_patches/call/rtp_video_sender.cc
@@ -598,7 +598,7 @@ EncodedImageCallback::Result RtpVideoSender::OnEncodedImage(
 //Stegozoa
 void RtpVideoSender::getSsrcs() {
     std::cout << "Ssrcs incoming" << std::endl;
-    for(int i = 0; i < rtp_streams_.size(); ++i)
+    for(size_t i = 0; i < rtp_streams_.size(); ++i)
     	std::cout << rtp_streams_[i].rtp_rtcp->SSRC() << " ";
     std::cout << std::endl;
 }
